+++ app/src/main/java/org/sopt/at/core/navigation/NavGraph.kt
@@ -6,13 +6,13 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import org.sopt.at.feature.main.history.HistoryScreen
-import org.sopt.at.feature.main.history.HistoryViewModel
-import org.sopt.at.feature.main.home.HomeScreen
-import org.sopt.at.feature.main.live.LiveScreen
-import org.sopt.at.feature.main.my.MyScreen
-import org.sopt.at.feature.main.search.SearchScreen
-import org.sopt.at.feature.main.shorts.ShortsScreen
+import org.sopt.at.feature.history.HistoryScreen
+import org.sopt.at.feature.history.HistoryViewModel
+import org.sopt.at.feature.home.HomeScreen
+import org.sopt.at.feature.live.LiveScreen
+import org.sopt.at.feature.my.MyScreen
+import org.sopt.at.feature.search.SearchScreen
+import org.sopt.at.feature.shorts.ShortsScreen
 import org.sopt.at.feature.signin.SignInScreen
 import org.sopt.at.feature.signup.SignUpScreen
 import org.sopt.at.feature.splash.SplashScreen
