+++ app/src/main/java/org/sopt/at/feature/shorts/ShortsScreen.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.feature.main.live
+package org.sopt.at.feature.shorts
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,17 +9,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 
 @Composable
-fun LiveScreen(modifier: Modifier = Modifier) {
+fun ShortsScreen(modifier: Modifier = Modifier) {
     Box(
         modifier = modifier.fillMaxSize(),
         contentAlignment = Alignment.Center
     ) {
-        Text("라이브 화면")
+        Text("숏츠 화면")
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-private fun PreviewLiveScreen() {
-    LiveScreen()
+private fun PreviewShortsScreen() {
+    ShortsScreen()
 }
